+++ src/render.h
@@ -44,7 +44,7 @@ void
 cmark_render_ascii(cmark_renderer *renderer, const char* s);
 
 void
-cmark_render_code_point(cmark_renderer *renderer, uint8_t c);
+cmark_render_code_point(cmark_renderer *renderer, uint32_t c);
 
 char*
 cmark_render(cmark_node *root,
